[JAM-8] StackOverflowError during the JaMoPPJDTParser::parseDirectory call
To reproduce this bug, we created a test case with a smallest input we could. This test runs without problems in the current version on the master branch. In contrast, this test case causes a StackOverflowError in the version on the jamopp branch. See: https://palladio-simulator.atlassian.net/browse/JAM-8