[JAM-10] Parameterized test for standalone use of JDT-based conversion of Java source code to EMF-based models.
Apart from JUint, no dependencies to other tools were newly introduced. For this purpose, the existing code was taken almost identically and copied into this test. This standalone implementation worked as expected in the first version we developed together. A comparable standalone call is also included in the original JaMoPP from DevBoost. This test generates corresponding EMF resources for the Java files in a project. JUinit passes the folder name of the projects <project> to be tested as a parameter. The EMF sources are then stored as XMI files in the folder ./standalone_output/<project>/. This ./standalone_output/ folder is not yet automatically deleted after test execution. Likewise, these saved EMF resources would have to be reloaded and then tested even further. However, since there are reproducible errors before, these should be fixed before reasonable testing can continue. As before, the projects do not need to be built. The projects are integrated with Git submodules, so with git submodule update --init all projects must be updated before test execution.